Changed the calculation to train over the last N weeks worth of data, starting at the last timestamp in the corrected dataset and going back N + 1 weeks through the second to last week.
This change also changes how timestamps are assigned to the predicted
values. The timestamps for the week that is left out of the training set
(the last week in the file) are used as the timestamps for the predicted
data timestamps. This is so the predictions can be passed into error
analysis easily.